requirements.txt
openai>=1.0.0
llama-index-core>=0.10.0
llama-index-embeddings-openai>=0.1.0
llama-hub>=0.0.79
llama-parse>=0.4.0
convex>=0.5.0
python-dotenv>=1.0.0
python-docx>=1.0.0
tqdm>=4.65.0
nest-asyncio>=1.5.0 